refactor(room): 신규 도메인 추가를 위한 구조 변경
